student_id and short_id

+++ smartscoreapp/views.py
@@ -2,32 +2,22 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth import authenticate, login, logout, get_user_model
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.contrib import messages
-from django.contrib.auth.models import User
 from .models import Class, Student, Exam, Question, Answer, ExamSet, StudentQuestion
-from .forms import ClassForm, StudentForm, ExamForm, ClassNameForm, EditStudentForm 
+from .forms import ClassForm, StudentForm, ExamForm, ClassNameForm, EditStudentForm, UserCreationWithEmailForm, AddStudentToExamForm, TestSetForm
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_http_methods
 from django.http import JsonResponse
-from .forms import UserCreationWithEmailForm 
 from django.core.exceptions import PermissionDenied
-from django.http import HttpResponseBadRequest
 from django.contrib.auth import update_session_auth_hash
 from django.contrib.auth.forms import PasswordChangeForm
-from random import shuffle
 import random
-from .forms import StudentForm, AddStudentToExamForm, TestSetForm
-from django.utils.crypto import get_random_string
-from django.db import models
-from .models import Student
 
-
-User = get_user_model() 
+User = get_user_model()
 
 def generate_student_code(exam_id):
     set_identifier = random.randint(0, 99)
     id_part = exam_id % 10000
     return f"{set_identifier:02d}-{id_part:04d}"
-
 
 def index(request):
     return render(request, 'index.html')
@@ -38,13 +28,10 @@ def login_view(request):
         if form.is_valid():
             username = form.cleaned_data.get('username')
             password = form.cleaned_data.get('password')
-            # Try to authenticate using the custom backend
             user = authenticate(request, username=username, password=password)
             if user is not None:
                 login(request, user)
                 messages.success(request, "Login successful!")
-                # Print the backend used
-                print(f"Backend used: {user.backend}")
                 return redirect('index')
             else:
                 messages.error(request, "Invalid username or password.")
@@ -61,7 +48,7 @@ def logout_view(request):
 
 def register_view(request):
     if request.method == 'POST':
-        form = UserCreationForm(request.POST)
+        form = UserCreationWithEmailForm(request.POST)
         if form.is_valid():
             form.save()
             username = form.cleaned_data.get('username')
@@ -69,13 +56,13 @@ def register_view(request):
             user = authenticate(username=username, password=password)
             login(request, user)
             messages.success(request, 'User registered successfully!')
-            return redirect('login')  # Redirect to the login page
+            return redirect('login')
         else:
             for field, errors in form.errors.items():
                 for error in errors:
                     messages.error(request, f"{field}: {error}")
     else:
-        form = UserCreationForm()
+        form = UserCreationWithEmailForm()
     
     return render(request, 'register.html', {'form': form})
 
@@ -101,14 +88,19 @@ def classes_view(request):
     
     context = {
         'classes': classes,
-        'form': form
+        'form': form,
+        'student_form': StudentForm()  # Add this line
     }
     return render(request, 'classes.html', context)
 
+@login_required
 def delete_class_view(request, class_id):
     class_instance = get_object_or_404(Class, id=class_id)
-    class_instance.delete()
-    messages.success(request, 'Class deleted successfully!')
+    if class_instance.user == request.user:
+        class_instance.delete()
+        messages.success(request, 'Class deleted successfully!')
+    else:
+        messages.error(request, 'You are not authorized to delete this class.')
     return redirect('classes')
 
 @login_required
@@ -126,16 +118,13 @@ def class_detail_view(request, class_id):
         if form.is_valid():
             student = form.save(commit=False)
             student.assigned_class = class_instance
-            print(f"Assigned class before save: {student.assigned_class}")  # Debug statement
             try:
                 student.save()
                 messages.success(request, 'Student added successfully!')
                 return redirect('class_detail', class_id=class_id)
             except Exception as e:
-                print(f"Error saving student: {e}")  # Debug statement
                 messages.error(request, f"Error saving student: {e}")
         else:
-            print(form.errors)  # Debug statement
             messages.error(request, 'Please correct the errors below.')
     else:
         form = StudentForm()
@@ -147,7 +136,7 @@ def class_detail_view(request, class_id):
     }
     return render(request, 'class_detail.html', context)
 
-
+@login_required
 def update_class_name_view(request, class_id):
     class_instance = get_object_or_404(Class, id=class_id)
 
@@ -168,7 +157,6 @@ def update_class_name_view(request, class_id):
     }
     return render(request, 'class_detail.html', context)
 
-
 @login_required
 def exams_view(request):
     user_instance = request.user
@@ -176,8 +164,11 @@ def exams_view(request):
     if request.method == 'POST':
         form = ExamForm(request.POST, user=user_instance)
         if form.is_valid():
-            form.save()
-            messages.success(request, 'Exam added successfully!')
+            exam = form.save(commit=False)
+            exam.save()
+            for i in range(3):  # Creating 3 sets for each exam
+                ExamSet.objects.create(exam=exam, set_number=i + 1)
+            messages.success(request, 'Exam and sets added successfully!')
             return redirect('exams')
         else:
             messages.error(request, 'There was an error adding the exam. Please check the form for errors.')
@@ -187,10 +178,6 @@ def exams_view(request):
     exams = Exam.objects.filter(class_assigned__user=user_instance)
     classes = Class.objects.filter(user=user_instance)
 
-    # Debug prints
-    print(f"Classes: {classes}")
-    print(f"Exams: {exams}")
-
     context = {
         'exams': exams,
         'form': form,
@@ -198,11 +185,14 @@ def exams_view(request):
     }
     return render(request, 'exams.html', context)
 
-
+@login_required
 def delete_exam(request, exam_id):
     exam = get_object_or_404(Exam, id=exam_id)
-    exam.delete()
-    messages.success(request, 'Exam deleted successfully.')
+    if exam.class_assigned.user == request.user:
+        exam.delete()
+        messages.success(request, 'Exam deleted successfully.')
+    else:
+        messages.error(request, 'You are not authorized to delete this exam.')
     return redirect('exams')
 
 @login_required
@@ -270,9 +260,9 @@ def edit_question_view(request, question_id):
         option_c = request.POST.get('option_c')
         option_d = request.POST.get('option_d')
         option_e = request.POST.get('option_e')
-        answer = request.POST.get('answer')
+        correct_answer = request.POST.get('answer')
 
-        if question_text and option_a and option_b and answer:
+        if question_text and option_a and option_b and correct_answer:
             # Update question
             question.question_text = question_text
             question.option_a = option_a
@@ -283,7 +273,7 @@ def edit_question_view(request, question_id):
             question.save()
 
             # Update answer
-            answer.answer = answer
+            answer.answer = correct_answer
             answer.save()
 
             messages.success(request, 'Question and answer updated successfully!')
@@ -299,7 +289,6 @@ def delete_question_view(request, question_id):
     exam_id = question.exam.id
 
     if request.method == 'POST':
-        # The associated answer will be automatically deleted due to the OneToOneField with on_delete=models.CASCADE
         question.delete()
         messages.success(request, 'Question and answer deleted successfully!')
         return redirect('exam_detail', exam_id=exam_id)
@@ -327,7 +316,6 @@ def select_questions_view(request, exam_id):
         selected_question_ids = request.POST.getlist('questions')
         selected_questions = Question.objects.filter(id__in=selected_question_ids)
         
-        # Clear existing questions and add selected questions
         exam.questions.clear()
         exam.questions.add(*selected_questions)
 
@@ -340,7 +328,7 @@ def select_questions_view(request, exam_id):
 def print_test_paper_view(request, exam_id):
     exam = get_object_or_404(Exam, id=exam_id)
     selected_questions = list(exam.questions.all())
-    random.shuffle(selected_questions)  # Randomize questions for each student
+    random.shuffle(selected_questions)
 
     return render(request, 'print_test_paper.html', {'exam': exam, 'selected_questions': selected_questions})
 
@@ -348,9 +336,8 @@ def print_test_paper_view(request, exam_id):
 def generate_test_paper_view(request, exam_id):
     exam = get_object_or_404(Exam, id=exam_id)
     selected_questions = list(exam.questions.all())
-    random.shuffle(selected_questions)  # Randomize questions for each student
+    random.shuffle(selected_questions)
 
-    # Prepare a list of questions with options for rendering
     questions_with_options = [{'question_text': question.question_text,
                                'options': [question.option_a, question.option_b, question.option_c, question.option_d, question.option_e]}
                               for question in selected_questions]
@@ -365,30 +352,24 @@ def process_scanned_papers_view(request):
         questions = list(exam.questions.all())
         total_questions = len(questions)
         
-        # Process scanned answers
         marks = []
         for question in questions:
-            answer_key = f'question{question.id}'  # Adjust as per your form structure
+            answer_key = f'question{question.id}'
             marked_option = request.POST.get(answer_key)
-            correct_option = question.answer  # Adjust based on your model design
+            correct_option = question.answer
 
             if marked_option is not None:
                 marks.append(1 if int(marked_option) == correct_option else 0)
             else:
-                marks.append(0)  # Handle case where no option is marked (if needed)
+                marks.append(0)
 
-        # Calculate score and other relevant metrics
         score = sum(marks)
         total_marks = len(marks)
 
-        # Optionally, save the scanned papers or marks for further processing
-        # Ensure your file paths and processing align with your requirements
-
         messages.success(request, f"Exam submitted successfully. Score: {score}/{total_marks}")
-        return redirect('exams')  # Redirect to exams page after processing
+        return redirect('exams')
 
-    return redirect('exams')  # Redirect to exams page if not a POST request or processing fails
-
+    return redirect('exams')
 
 def generate_answers_list(exam):
     selected_questions = exam.questions.all()
@@ -451,11 +432,9 @@ def exam_detail_view(request, exam_id):
         else:
             messages.error(request, "Question text is required!")
 
-    # Retrieve questions related to the exam
     questions = exam.question_set.all()
 
     return render(request, 'exam_detail.html', {'exam': exam, 'questions': questions})
-
 
 def students_view(request):
     students = Student.objects.all()
@@ -464,45 +443,39 @@ def students_view(request):
 @login_required
 def add_student_view(request, class_id):
     class_instance = get_object_or_404(Class, id=class_id)
-    print(f"Class instance: {class_instance}")  # Debug statement
 
     if request.method == 'POST':
         form = StudentForm(request.POST)
         if form.is_valid():
             student = form.save(commit=False)
             student.assigned_class = class_instance
-            print(f"Assigned class before save: {student.assigned_class}")  # Debug statement
             try:
                 student.save()
-                print(f"Student saved with assigned_class: {student.assigned_class}")  # Debug statement
-                messages.success(request, 'Student added successfully!')
-                return redirect('class_detail', class_id=class_id)
+                messages.success(request, f'Student added successfully! Student ID: {student.short_id}')
+                return redirect('classes')
             except Exception as e:
-                print(f"Error saving student: {e}")  # Debug statement
                 messages.error(request, f"Error saving student: {e}")
         else:
-            print(form.errors)  # Debug statement
             messages.error(request, 'Please correct the errors below.')
-    else:
-        form = StudentForm()
-
-    return render(request, 'add_student.html', {'form': form, 'class': class_instance})
-
+    
+    return redirect('classes')
 
 @login_required
 def edit_student(request, student_id):
     student = get_object_or_404(Student, pk=student_id)
     
     if request.method == 'POST':
-        student.name = request.POST.get('name')
-        student.save()
-        # Optionally add success message
-        messages.success(request, 'Student updated successfully.')
-        return redirect('class_detail', class_id=student.assigned_class.id)  # Redirect to appropriate page
+        form = EditStudentForm(request.POST, instance=student)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Student updated successfully.')
+            return redirect('class_detail', class_id=student.assigned_class.id)
+        else:
+            messages.error(request, 'Please correct the errors below.')
     
-    # Handle other HTTP methods or render form for GET request
-    return render(request, 'edit_student.html', {'student': student})
+    return render(request, 'edit_student.html', {'form': form, 'student': student})
 
+@login_required
 def delete_student(request, student_id):
     student = get_object_or_404(Student, pk=student_id)
 
@@ -512,10 +485,7 @@ def delete_student(request, student_id):
         messages.success(request, 'Student deleted successfully.')
         return redirect('class_detail', class_id=assigned_class_id)
 
-    # Handle other HTTP methods if necessary
-
     return redirect('class_detail', class_id=student.assigned_class.id)
-
 
 @login_required
 def add_student_to_exam_view(request, exam_id):
@@ -547,7 +517,7 @@ def add_class_view(request):
         form = ClassForm(request.POST)
         if form.is_valid():
             new_class = form.save(commit=False)
-            new_class.user = request.user  # Directly assign request.user
+            new_class.user = request.user
             new_class.save()
             messages.success(request, 'Class added successfully!')
             return redirect('classes')
@@ -580,8 +550,6 @@ def settings_view(request):
 @require_http_methods(['POST'])
 def ajax_get_students(request):
     class_id = request.POST.get('class_id')
-    students = Student.objects.filter(assigned_class_id=class_id).values('id', 'name')
+    students = Student.objects.filter(assigned_class_id=class_id).values('id', 'first_name', 'last_name', 'student_id', 'short_id')
     data = list(students)
     return JsonResponse(data, safe=False)
-
-
